+++ py-polars/polars/__init__.py
@@ -127,7 +127,6 @@ from polars.functions.lazy import date_ as date
 from polars.functions.lazy import datetime_ as datetime
 from polars.functions.lazy import list_ as list
 from polars.functions.whenthen import when
-from polars.internals.io import read_ipc_schema, read_parquet_schema
 from polars.io import (
     read_avro,
     read_csv,
@@ -136,9 +135,11 @@ from polars.io import (
     read_delta,
     read_excel,
     read_ipc,
+    read_ipc_schema,
     read_json,
     read_ndjson,
     read_parquet,
+    read_parquet_schema,
     read_sql,
     scan_csv,
     scan_delta,
