+++ py-polars/polars/io/ipc/__init__.py
@@ -1,0 +1,7 @@
+from polars.io.ipc.functions import read_ipc, read_ipc_schema, scan_ipc
+
+__all__ = [
+    "read_ipc",
+    "read_ipc_schema",
+    "scan_ipc",
+]
