+++ py-polars/polars/io/pyarrow_dataset/__init__.py
@@ -1,0 +1,6 @@
+from polars.io.pyarrow_dataset.functions import scan_ds, scan_pyarrow_dataset
+
+__all__ = [
+    "scan_pyarrow_dataset",
+    "scan_ds",
+]
